Hao dataset download and processor
This script requires the preHaoDownloader.py to be in the same directory. It uses the python script to download all fastq files in the Hao dataset, and concatonates them into one large dataset. Then it processes this dataset, aligning it to the hg19 human reference genome to produce a bam file. At different stages during the processes quality assesments are provided to check all steps.